lib: Move toID to its own file
This allows for some memory savings by decoupling its import from dex-data
