+++ sim/dex-abilities.ts
@@ -1,5 +1,6 @@
 import type { PokemonEventMethods, ConditionData } from './dex-conditions';
-import { assignMissingFields, BasicEffect, toID } from './dex-data';
+import { assignMissingFields, BasicEffect } from './dex-data';
+import { toID } from '../lib/to-id';
 import { Utils } from '../lib/utils';
 
 interface AbilityEventMethods {
