+++ sim/dex-species.ts
@@ -1,4 +1,5 @@
-import { assignMissingFields, BasicEffect, toID } from './dex-data';
+import { assignMissingFields, BasicEffect } from './dex-data';
+import { toID } from '../lib/to-id';
 import { Utils } from '../lib/utils';
 import { isDeepStrictEqual } from 'node:util';
